@@ -17,22 +17,21 @@
     "authors": [
         {
             "name": "Vincent Klaiber",
-            "email": "hello@doubledip.se",
-            "homepage": "https://doubledip.se"
+            "email": "hello@doubledip.se"
         }
     ],
     "require": {
-        "php": "^7.2",
-        "graham-campbell/manager": "^4.3",
+        "php": "^7.4",
+        "graham-campbell/manager": "^4.4",
         "hashids/hashids": "^4.0",
-        "illuminate/contracts": "^6.0",
-        "illuminate/support": "^6.0"
+        "illuminate/contracts": "^7.0",
+        "illuminate/support": "^7.0"
     },
     "require-dev": {
-        "graham-campbell/analyzer": "^2.3",
-        "graham-campbell/testbench": "^5.3",
-        "mockery/mockery": "^1.2",
-        "phpunit/phpunit": "^8.0",
+        "graham-campbell/analyzer": "^2.4",
+        "graham-campbell/testbench": "^5.4",
+        "mockery/mockery": "^1.3",
+        "phpunit/phpunit": "^8.5",
         "squizlabs/php_codesniffer": "^3.5"
     },
     "config": {
@@ -40,7 +39,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "7.0-dev"
+            "dev-master": "8.0-dev"
         },
         "laravel": {
             "providers": [
