@@ -1,12 +1,12 @@
 <?php
 
-/*
- * This file is part of Laravel Hashids.
- *
- * (c) Vincent Klaiber <hello@doubledip.se>
+/**
+ * Copyright (c) Vincent Klaiber.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
+ *
+ * @see https://github.com/vinkla/laravel-hashids
  */
 
 declare(strict_types=1);
@@ -16,10 +16,6 @@ namespace Vinkla\Hashids\Facades;
 use Illuminate\Support\Facades\Facade;
 
 /**
- * This is the Hashids facade class.
- *
- * @author Vincent Klaiber <hello@doubledip.se>
- *
  * @method static string encode(mixed ...$numbers) Encode parameters to generate a hash.
  * @method static array decode(string $hash) Decode a hash to the original parameter values.
  * @method static string encodeHex(string $str) Encode hexadecimal values and generate a hash string.
@@ -27,11 +23,6 @@ use Illuminate\Support\Facades\Facade;
  */
 class Hashids extends Facade
 {
-    /**
-     * Get the registered name of the component.
-     *
-     * @return string
-     */
     protected static function getFacadeAccessor(): string
     {
         return 'hashids';
