@@ -1,12 +1,12 @@
 <?php
 
-/*
- * This file is part of Laravel Hashids.
- *
- * (c) Vincent Klaiber <hello@doubledip.se>
+/**
+ * Copyright (c) Vincent Klaiber.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
+ *
+ * @see https://github.com/vinkla/laravel-hashids
  */
 
 declare(strict_types=1);
@@ -17,20 +17,10 @@ use GrahamCampbell\Analyzer\AnalysisTrait;
 use Laravel\Lumen\Application;
 use PHPUnit\Framework\TestCase;
 
-/**
- * This is the analysis test class.
- *
- * @author Vincent Klaiber <hello@doubledip.se>
- */
 class AnalysisTest extends TestCase
 {
     use AnalysisTrait;
 
-    /**
-     * Get the code paths to analyze.
-     *
-     * @return string[]
-     */
     protected function getPaths()
     {
         return [
@@ -40,11 +30,6 @@ class AnalysisTest extends TestCase
         ];
     }
 
-    /**
-     * Get the classes to ignore not existing.
-     *
-     * @return string[]
-     */
     protected function getIgnored()
     {
         return [Application::class];
