@@ -1,12 +1,12 @@
 <?php
 
-/*
- * This file is part of Laravel Hashids.
- *
- * (c) Vincent Klaiber <hello@doubledip.se>
+/**
+ * Copyright (c) Vincent Klaiber.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
+ *
+ * @see https://github.com/vinkla/laravel-hashids
  */
 
 declare(strict_types=1);
@@ -18,40 +18,20 @@ use Vinkla\Hashids\Facades\Hashids;
 use Vinkla\Hashids\HashidsManager;
 use Vinkla\Tests\Hashids\AbstractTestCase;
 
-/**
- * This is the Hashids facade test class.
- *
- * @author Vincent Klaiber <hello@doubledip.se>
- */
 class HashidsTest extends AbstractTestCase
 {
     use FacadeTrait;
 
-    /**
-     * Get the facade accessor.
-     *
-     * @return string
-     */
     protected function getFacadeAccessor()
     {
         return 'hashids';
     }
 
-    /**
-     * Get the facade class.
-     *
-     * @return string
-     */
     protected function getFacadeClass()
     {
         return Hashids::class;
     }
 
-    /**
-     * Get the facade root.
-     *
-     * @return string
-     */
     protected function getFacadeRoot()
     {
         return HashidsManager::class;
